+++ include/ast/node/float_list.h
@@ -11,7 +11,7 @@ class Float_list : public Node {
         Float_list(std::string str, U64 hash, int num_floats_in_list):
             Node(str, hash)
         {
-            constraint = std::make_optional<Node_constraint>(Common::float_literal, num_floats_in_list);
+            constraint = std::make_optional<Node_constraint>(Common::float_literal, size_t(num_floats_in_list));
         }
 
     private:
