+++ include/ast/node/resource_list.h
@@ -6,7 +6,7 @@
 class Qubit_list : public Node {
 
     public:
-        Qubit_list(std::string str, U64 hash, int num_qubits_in_list):
+        Qubit_list(std::string str, U64 hash, size_t num_qubits_in_list):
             Node(str, hash)
         {
             constraint = std::make_optional<Node_constraint>(Common::qubit, num_qubits_in_list);
@@ -19,7 +19,7 @@ class Qubit_list : public Node {
 class Bit_list : public Node {
 
     public:
-        Bit_list(std::string str, U64 hash, int num_bits_in_list):
+        Bit_list(std::string str, U64 hash, size_t num_bits_in_list):
             Node(str, hash)
         {
             constraint = std::make_optional<Node_constraint>(Common::bit, num_bits_in_list);
